==> notebooks/part_01.ETL/part_01.creating_web_scrapper.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup as Bs 
import re 
import lxml
import time 

def get_html(url):
    "return html"
    time.sleep(1)
    res = requests.get(url)
    res.raise_for_status() 
    return res.content

def get_soup(html):
    "return organized html"
    return Bs(html, "html.parser")

def get_href(soup):
    "return addresses to second web"
    return [i if i.startswith("/") else "/" + i for i in [i["href"] for i in soup.find_all("a")[:-1]]]

def generate_url(href):
    return [f"https://planecrashinfo.com{i}" for i in href]

# int(test_soup.find("title").text)
def generate_url_outside(href):
    return [f"https://planecrashinfo.com/{int(i.split('-')[0][1:])}{i}" for i in href]

def decompress_list(lis):
    return [m for i in lis for m in i]

def extract_data(soup_table):
    accident_data = {}
    # Loop through rows and extract key-value pairs
    for row in soup_table[1:]:  # Skip header rows
        key_cell = row.find('td', align='right')
        value_cell = row.find('td', align='left')
        if key_cell and value_cell:
            key = key_cell.text.strip().replace(':', '')  # Remove colon and trailing spaces
            value = value_cell.text.strip()
            accident_data[key] = value
    return accident_data

def first_page_pass(url):
    html = get_html(url)
    soup = get_soup(html)
    href = get_href(soup)
    href_url = generate_url(href)
    return href_url

def second_page_pass(href_url):
    return [generate_url_outside(i) for i in [get_href(i) for i in [get_soup(i) for i in [get_html(i) for i in href_url]]]]

def third_final_pass(good_url):
    return [extract_data(i) for i in [get_soup(j).find_all("tr") for j in [get_html(i) for i in good_url ]]]

def save_csv(my_list,name):
    j = pd.DataFrame(my_list)
    j.to_csv(f"data/{name}", index=False)

url = "https://www.planecrashinfo.com/database.htm"

href_url = first_page_pass(url)
good_url = second_page_pass(href_url)
save_csv(good_url,"urls/"+"list_of_list_urls.csv")
k = decompress_list(good_url)
save_csv(k,"urls/"+"list_of_urls.csv")

(len(good_url),len(k))

good_url[-20:]

for i in good_url[-20:]:
    time.sleep(3)
    new_list = third_final_pass(i)
    save_csv(new_list,"data/bronze/" + i[0].split("/")[-2] + ".csv")
