==> notebooks/part_01.ETL/part_02.preliminary_clean_and_join.ipynb <==
### Preliminary data cleaning:
- Combine datasets into one 
- Clean data names 
- Make Aboard into dictionaries 
- Replace "?" with nan 
- add Decades column 
- Fix date time column
- Define data types 
- Assumptions "?" == 0

import pandas as pd 
from functions import extraction_and_cleaning

extraction_and_cleaning??

df = extraction_and_cleaning.read_bronze()
df.head(5)

First lets fix the column names

new_names = []
for item in df.columns:
    item = item.lower()
    item = item.strip()
    item = item.replace("\n"," ")
    item = item.replace("/"," ")
    item = item.replace("#","num")
    item = item.split()
    item = "_".join(item)
    new_names.append(item.title())

df.columns = new_names

def clean_aboard(x):
    x = x.replace("(","").replace(")","").split()
    a, b, c = x[0], x[1].split(":")[-1], x[2].split(":")[-1]
    try: 
        return {"total":int(a),"passengers":int(b),"crew":int(c)}
    except:
        if a == "?":
            a = float("nan")
        else:
            a = int(a)
        if b == "?":
            b = float("nan")
        else: 
            b = int(b)
        if c == "?":
            c = float("nan")
        else: 
            c = int(c)
    return {"total":a,"passengers":b,"crew":c}

df.Aboard = df.Aboard.apply(clean_aboard)
df.Fatalities = df.Fatalities.apply(clean_aboard)
df.head(3)

# Example getting flights with total passengers  + crew greater than 500 
df[df.Aboard.apply(lambda x: x["total"] > 500)]

df[df == "?"] = float("nan")
df["Date_2"] = pd.to_datetime(df.Date)
df.head(3)

df.info()

df["Decade"] = df.Date_2.dt.year // 10 * 10 

df = df[['Decade', 'Date','Time', 'Location', 'Operator', 'Flight_Num', 'Route',
       'Ac_Type', 'Registration', 'Cn_Ln', 'Aboard', 'Fatalities', 'Ground',
       'Summary']]

df[df.Aboard.apply(lambda x: isinstance(x['total'],float))]

### Drop nan for Fatalities

print(df.shape)
df = df.drop(df[df.Aboard.apply(lambda x: isinstance(x["total"],float))].index)

def remove_total(x):
    return x["total"]

df["Aboard_Total"]  = df.Aboard.apply(remove_total)
df["Fatalities_Total"] = df.Fatalities.apply(remove_total)
df["Operator_vs_Military"] = df.Operator.replace(r'\b.*Military.*\b', 'Military', regex=True)
df.shape 

sum(df.Location.apply(lambda x: not isinstance(x,float))), sum(df.Location.apply(lambda x: isinstance(x,float)))

df.to_json('data/silver/1920_2024.json')
