==> notebooks/part_01.ETL/part_04.Geolocation_Geopy.SanityCheck.ipynb <==
#### After we got the data from Geopy 
- Check column length if we managed to get all the data and the answer is no
- About 97 records have null in the Country_location column 
- we should have 5015 records

import pandas as pd

Read in the 1920_2024_geopy and 

df = pd.read_json("data/silver/1920_2024_geopy.json")
df[:5]

len(df)

sum(df.Country_location.apply(lambda x: isinstance(x,dict))), sum(df.Country_location.apply(lambda x: not isinstance(x,dict)))

df[df.Country_location.apply(lambda x: not isinstance(x,dict))]

We are going to save the bad data for further processing and eda 

df[df.Country_location.apply(lambda x: isinstance(x,str))].to_csv("data/bad_data/unknown_country_location.csv")

### Unfortunately After the geopy we have 97 data points that weren't able to be locate
- We have some additional cleaning maybe some formatting will do that next
