==> notebooks/part_01.ETL/part_05.Geolocation_Geopy.Pull_Data_Fixing_bad_data.ipynb <==
## EDA clean AC tpye 
- Just get the name without the plane type

import pandas as pd
from geopy.geocoders import Nominatim

df = pd.read_csv("data/bad_data/unknown_country_location.csv")

for i in df.Country_location:
    print(i)

df["Country_location"] = df.Location

df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Off Northern Germany", "Germany"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Mount Phou-Lassy, French Indo-China", "Luang Prabang, Laos"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Bay ", "Bay, "))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Alakska", "Alaska"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Ocean ","Ocean, "))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Island ","Island, "))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Mt. Helmos. ","Mt. Helmos., "))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Near Lete Pass","Lete, Nepal"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("South Kazakistan","Kazakistan"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Sao Gabriel de Cachoeria","Sao Gabriel da Cachoeira"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("Mendotta, Minnisota","Mendota, Minnesota"))
df["Country_location"] = df.Country_location.apply(lambda x: x.replace("French Cameroons","Cameroon"))

for i in df.Country_location:
    print(i)

def remove_bad_words(inputString,ls=["Off ", "Over ", "Near ", "Air ", "Base ", "the ", "Centeral "]):
    for i in ls:
        if i.lower() in inputString.lower():
            inputString = inputString.replace(i,"")
    return inputString

df["Country_location"] = df.Country_location.apply(lambda x: remove_bad_words(x))

for i in df.Country_location:
    print(i)

geolocator = Nominatim(user_agent="specify_your_app_name")

def country_of_origin(x):
    try:
        try:
            try:
                return geolocator.geocode(x, language='en').raw
            except: 
                return geolocator.geocode(x.split(",")[0], language='en').raw
        except:
            return geolocator.geocode(x.split(",")[-1], language='en').raw
    except: float("nan")

df

df["Country_location"] = df.Country_location.apply(country_of_origin)

df[df["Country_location"].apply(lambda x: not isinstance(x,dict))]

df.to_csv("data/bad_data/cleaned_bad_data/cleaned_global_data.csv")

df[df["Country_location"].apply(lambda x: not isinstance(x,dict))].to_csv("data/bad_data/cleaned_bad_data/un-cleanable_cleaned_global_data_.csv")

Only 2 bad record left that couldn't be identified
