==> notebooks/part_01.ETL/part_06.Geolocation_Geopy.Combine.ipynb <==
import pandas as pd
import ast 

df = pd.read_json("data/silver/1920_2024_geopy.json")
df_2 = pd.read_csv("data/bad_data/cleaned_bad_data/cleaned_global_data.csv")

df[:2]

df[df.Country_location.apply(lambda x: isinstance(x,str))][:2]

df[df.Country_location.apply(lambda x: isinstance(x,str))].shape

df_2[:1]

df_2 = df_2[df_2.columns[2:]]
df_2[:2]

df[df.Country_location.apply(lambda x: isinstance(x,str))].iloc[:,3:5]

df_2.iloc[:,3:5]

sum(df[df.Country_location.apply(lambda x: isinstance(x,str))].iloc[:,3:5].values == df_2.iloc[:,3:5].values)

k = df.loc[df['Country_location'].apply(lambda x: isinstance(x, str))]
k[(k["Location"] !=  "175 miles off the Egyptian coast") & (k["Location"] != "Djiring, French Indo-China")].shape

df_2[df_2['Country_location'].apply(lambda x: isinstance(x,str))].shape

a = k[(k["Location"] !=  "175 miles off the Egyptian coast") & (k["Location"] != "Djiring, French Indo-China")].index

a 

df_2[df_2['Country_location'].apply(lambda x: isinstance(x, str))]['Country_location'].apply(ast.literal_eval).values[0]

df.loc[a, "Country_location"] = df_2[df_2['Country_location'].apply(lambda x: isinstance(x, str))]['Country_location'].apply(ast.literal_eval).values

df.loc[a]

df["Date"] = df["Date"].astype(str)

df.to_json("data/silver/1920_2024_geopy_fixed.json")
