==> notebooks/part_01.ETL/part_09.Geolocation_Geopy.USA_state.ipynb <==
import pandas as pd

Read in the 1920_2024_geopy and 

df = pd.read_csv("data/ready_for_plotting/data_plot_map.csv")
df = df[df.Code == "USA"]
df[:3]

df_state_codes = pd.read_json("data/state_codes.json")
df_state_codes[:3]

def unpack_locations(x,a=-2):
    "Check if is numeric"
    return ",".join([i for i in x['display_name'].split(",") if i.strip().isnumeric()==False][a:]).strip()

def get_country_from_dict(df):
    df["Country_Geopy"] = df.Country_location.apply(unpack_locations)
    df = df.set_index("Country_Geopy")
    return df

def check_and_compare(dfContryCode,seriesIndex):
    return dfContryCode.loc[seriesIndex]["code"]

def get_code(country_code,x):
    return  check_and_compare(country_code, x)

state_code = get_country_from_dict(df_state_codes)
state_code[:3]

state_code.rename(index={'District of Columbia, United States':'Washington, United States'},inplace=True)
state_code.loc['District of Columbia, United States'] = ["DC", 'District of Columbia', float("nan")]  

state_code

df[df.Country_Geopy.apply(lambda x: x =="United States")]

df[df.Country_Geopy.apply(lambda x: x not in state_code.index)]

df["State_Code"] = df.Country_Geopy.apply(lambda x: x != "United States" and x != "Wake Island, United States" and get_code(state_code,x))

df[:5]

df.shape

df.to_csv("data/ready_for_plotting/usa_data.csv", index=False)
